fix: refactor code
- move main to pokeGo
- create helper function
- rename InterfaceRepository into GlobalInterface
- check for each layer connection

+++ internal/domain/pokemon.go
@@ -4,7 +4,7 @@ type Pokemon struct {
 	ID             int
 	Name           string
 	Type           string
-	CatchRate      int
+	CatchRate      float64
 	IsRare         bool
 	RegisteredDate string
 }
